requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
pydantic==2.9.2
python-multipart==0.0.9
httpx==0.27.2
Pillow==10.4.0
ImageHash==4.3.1
playwright==1.55.0
numpy==2.1.1
python-dateutil==2.8.2
opencv-python==4.8.1.78
pytesseract==0.3.10
openai==1.12.0
requests==2.31.0
psycopg[binary]==3.2.10
python-dotenv==1.0.0
